refactor: extract shared panel shell and split diagnostics page components
Addresses the clean-code findings in frontend/src/pages/diagnostics.tsx:
four hand-copied panel shells, a magic status-dot size, a duplicated
warning-alert class string, and two over-long functions in a 453-line file.

- Add a shared <Panel> shell (components/diagnostics/panel.tsx) owning the
  section/cardVariants wrapper and h2 heading the four panels repeated, plus
  a PANEL_WARNING_CLASS constant for the warning banner the telemetry and
  logging panels both inlined.
- Move the service row, the four panels, the drop-counter row, and
  mergeServices into components/diagnostics/, one component per file,
  matching the components/app-detail/ convention. DiagServiceRow loses its
  now-redundant prefix and its status/phase/retry annotations move to a
  ServiceRowMeta child (62 -> 36 lines).
- Pull the page's fetch/merge/derivation into useDiagnosticsData(), leaving
  DiagnosticsPage as markup (89 -> 55 lines); the page file drops to 147.
- Add a co-located unit test for the extracted mergeServices logic.

The service row's status dot changes size 8 -> STATUS_DOT_SIZE (10). That is
an intentional normalization, not a copy-paste: the constant is documented as
the size for table cells and list rows, the boot-issues list in the same file
already used it, and the smaller BADGE_STATUS_DOT_SIZE (8) is for pill/badge
contexts this row is not. It is the only rendered change in this commit.

Closes #1667
